added labels for modal accessiblity, django test, and fix linting
